Added zero mean shifting, fixed hessian estimation
Applied zero mean shifting at the end of each iteration with an update
on the supplied image. Calcuating the Hessian matrix approximation by
using only the jacobian part and not the whole gradient calculation
which computes a multiplication with the residual. Added an option for
solution contraint mask which is applied to the newly transformed image
at the end of each update. Finally remove the old code for
transformation estimation and cleaned up the imports.